Corrected for spelling and RegEx validation
This template contained several misspellings of "SMTP" and--I'm not entirely sure what the exact cause of this was, as I could not repeat the errors on templates with similar formatting--the RegEx rules would persistently fail for the SMTP boolean settings. I tested across several browsers and encountered the same problem. After wrapping the default values with ' ' they stopped failing the RegEx check. This pull request should be paired with another with similar changes in the GH Pages template.